+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * updated to Gradle 4.4
   * bundle license and dependency information
     ([JENKINS-27729](https://issues.jenkins-ci.org/browse/JENKINS-27729))
+  * added `jenkinsServer` configuration that can be used to install extra plugins for the `server` task
 
 ## 0.25.0 (2018-01-15)